backend/requirements.txt
Flask[async]==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
pymongo==4.6.1
motor==3.3.2
python-dotenv==1.0.0
Werkzeug==3.0.1
email-validator==2.1.0
python-dateutil==2.8.2
asgiref==3.7.2
